pyproject.toml
[project]
name = "coder-buddy"
version = "0.1.0"
description = "AI Engineering Project Planner - Automatically plans, architects, and implements complete engineering projects from natural language prompts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.0",
    "langgraph>=0.6.3",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []